Harmonize use of int (32 or 64) and stop using deprecated numpy aliases
- avoid using deprecated np.int and (-> np.int_) and np.bool (->
  np.bool_)
- use np.int_ instead of np.int32 or np.int64. s dtype. This is
  plateform dependent (correspond to C "long") and is what numpy
  transforms dtype=int into
- use numba.types.int_ instead of numba.types.int32 (same idea as
  np.int_)

+++ discrete_optimization/generic_tools/mutations/mixed_mutation.py
@@ -21,7 +21,7 @@ class BasicPortfolioMutation(Mutation):
         if isinstance(self.weight_mutation, list):
             self.weight_mutation = np.array(self.weight_mutation)
         self.weight_mutation = self.weight_mutation / np.sum(self.weight_mutation)
-        self.index_np = np.array(range(len(self.list_mutation)), dtype=np.int32)
+        self.index_np = np.array(range(len(self.list_mutation)), dtype=np.int_)
 
     def mutate(self, solution: Solution) -> Tuple[Solution, LocalMove]:
         choice = np.random.choice(self.index_np, size=1, p=self.weight_mutation)[0]
@@ -45,7 +45,7 @@ class BasicPortfolioMutationTrack(Mutation):
         if isinstance(self.weight_mutation, list):
             self.weight_mutation = np.array(self.weight_mutation)
         self.weight_mutation = self.weight_mutation / np.sum(self.weight_mutation)
-        self.index_np = np.array(range(len(self.list_mutation)), dtype=np.int32)
+        self.index_np = np.array(range(len(self.list_mutation)), dtype=np.int_)
 
     def mutate(self, solution: Solution) -> Tuple[Solution, LocalMove]:
         choice = np.random.choice(self.index_np, size=1, p=self.weight_mutation)[0]
